Adding name of cheese on hover
Totally not in my brief but just a trial of getting the name of the
cheese to appear when you hover over the image - in case people can’t
identify which cheese it is from the photo. Happy for people to get rid
of it if they think it’s a bit naff.