Sorting out text on hover as had become misaligned!
Hi guys, the text on the hover had got a bit messed up so I have
realigned it all. I know the font is a little small but pls can we
leave it this size as now it its nicely in the box and I don’t want it
to go all wonky again! Thanks! Ruth (on my dying laptop….)